correct output

+++ C++/Striping.cpp
@@ -8,8 +8,8 @@ average(average), reflection(z, dc)
 
 unsigned char Striping::calculate_bw()
 {
-  return static_cast<unsigned char>((static_cast<double>(maxColorValue) - 1.0) - (100.0 *
-  average) * reflection.calculate()); // explicit casting
+  return static_cast<unsigned char>(((maxColorValue - 1) - (100.0 *
+  average)) * reflection.calculate()); // explicit casting
 }
 
 unsigned char Striping::calculate_r()
